Refactor codebase and leveraged global state
