+++ src/main.js
@@ -1,6 +1,7 @@
 import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
+import { store, actions } from "./store";
 
 // Toast
 import Toast from "vue-toastification";
@@ -10,6 +11,8 @@ import "vue-toastification/dist/index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 createApp(App)
+  .provide("store", store)
+  .provide("actions", actions)
   .use(router)
   .use(Toast, {
     transition: "Vue-Toastification__bounce",
